Add files via upload
The SQL project involves managing and analyzing sales data for a car dealership company. In this project, the data analyst creates SQL queries to extract relevant information about car sales. The queries aim to provide valuable insights and reports that are then presented to the manager. The project focuses on leveraging SQL to efficiently retrieve, analyze, and present sales-related information for informed decision-making within the car dealership.